refactor register page code
