+++ src/Components/forms/RegisterForm.jsx
@@ -1,163 +1,104 @@
 import React from 'react';
-import axios from 'axios';
+import { PropTypes } from 'prop-types';
 import { Link } from 'react-router-dom';
-import { toast } from 'react-toastify';
-import Authservice from '../Auth/AuthService';
-import { baseURL as url } from '../../utils/Config';
 import InputField from '../inputs/InputField';
 import Button from '../inputs/Button';
 
-class RegisterForm extends React.Component {
-	constructor(props) {
-		super(props);
-		this.state = {
-			username: '',
-			email: '',
-			password: '',
-			cpassword: '',
-			first_name: '',
-			last_name: '',
-			errors: {},
-		};
-		this.auth = new Authservice();
-		this.handleChange = this.handleChange.bind(this);
-		this.confirmPass = this.confirmPass.bind(this);
-		this.handleSubmit = this.handleSubmit.bind(this);
-	}
-	componentDidMount() {
-		if (this.auth.isLoggedIn()) {
-			//this.props.history.replace('/');
-		}
-	}
-
-	handleChange(e) {
-		this.setState({
-			[e.target.name]: e.target.value,
-		});
-	}
-	confirmPass(e) {
-		this.setState({
-			cpassword: e.target.value,
-		});
-	}
-	handleSubmit(event) {
-		event.preventDefault();
-		const user = {
-			username: this.state.username,
-			email: this.state.email,
-			password: this.state.password,
-			first_name: this.state.first_name,
-			last_name: this.state.last_name,
-		};
-
-		axios({
-			method: 'post',
-			url: 'auth/register',
-			data: user,
-			baseURL: url,
-			responseType: 'json',
-			headers: {
-				'Content-Type': 'application/json'
-			},
-		}).then((response) => {
-			if (response.status >= 200 && response.status < 300) {
-				this.props.history.replace('/login');
-				toast.success(() =>
-					(<div>
-						<h3>Success</h3>
-						<p>Your Account was successfully Created<br />
-							Login now with your username and password
-						</p>
-					</div>)
-				);
-			}
-		}).catch((error) => {
-			if (error.response !== undefined) {
-				this.setState({ errors: error.response.data['Errors'] });
-			}
-			else {
-				toast.error(() => <div>
-					<h3>Opps!!</h3>
-					<p>Sorry! Something went wrong. If the problem persist, contact support</p>
-				</div>);
-			}
-		});
-	}
-	render() {
-		return (
-			<div className="container wrapper">
-				<div className="main-center">
-					<div className="card">
-						<div className="card-body">
-							<h3>Create Account</h3>
-							<form className="form" noValidate onSubmit={this.handleSubmit}>
-								<InputField
-									type='text'
-									name='username'
-									label='Username'
-									onChange={this.handleChange}
-									placeholder='username'
-									value={this.state.username}
-									error={this.state.errors['username']}
-								/>
-								<InputField
-									type='email'
-									name='email'
-									label='Email'
-									onChange={this.handleChange}
-									placeholder='email'
-									value={this.state.email}
-									error={this.state.errors['email']}
-								/>
-								<InputField
-									type='text'
-									name='first_name'
-									label='First Name'
-									onChange={this.handleChange}
-									placeholder='First Name'
-									value={this.state.first_name}
-									error={this.state.errors['first_name']}
-								/>
-								<InputField
-									type='text'
-									name='last_name'
-									label='Last Name'
-									onChange={this.handleChange}
-									placeholder='Last Name'
-									value={this.state.last_name}
-									error={this.state.errors['last_name']}
-								/>
-								<InputField
-									type='password'
-									name='password'
-									label='Password'
-									onChange={this.handleChange}
-									placeholder='password'
-									value={this.state.password}
-									error={this.state.errors['password']}
-								/>
-								<InputField
-									type='password'
-									name='confirm_password'
-									label='Confirm Password'
-									onChange={this.confirmPass}
-									placeholder='confirm password'
-									value={this.state.cpassword}
-									error={(this.state.cpassword !== this.state.password) && 'Password mismatch'}
-								/>
-								<Button
-									type="submit"
-									className="btn btn-lg btn-success btn-block btn-signin"
-									text="Register"
-								/>
-								<p className="message">Alredy registered? <Link to="/login">Signin</Link></p>
-							</form>
-						</div>
+const RegisterForm = ({ 
+	handleChange, 
+	handleSubmit,
+	confirmPass,
+	email,
+	cpassword, 
+	first_name, 
+	last_name, 
+	errors, 
+	username, 
+	password 
+}) => {
+	return (
+		<div className="container wrapper">
+			<div className="main-center">
+				<div className="card">
+					<div className="card-body">
+						<h3>Create Account</h3>
+						<form className="form" noValidate onSubmit={handleSubmit}>
+							<InputField
+								type='text'
+								name='username'
+								label='Username'
+								onChange={handleChange}
+								placeholder='username'
+								value={username}
+								error={errors['username']}
+							/>
+							<InputField
+								type='email'
+								name='email'
+								label='Email'
+								onChange={handleChange}
+								placeholder='email'
+								value={email}
+								error={errors['email']}
+							/>
+							<InputField
+								type='text'
+								name='first_name'
+								label='First Name'
+								onChange={handleChange}
+								placeholder='First Name'
+								value={first_name}
+								error={errors['first_name']}
+							/>
+							<InputField
+								type='text'
+								name='last_name'
+								label='Last Name'
+								onChange={handleChange}
+								placeholder='Last Name'
+								value={last_name}
+								error={errors['last_name']}
+							/>
+							<InputField
+								type='password'
+								name='password'
+								label='Password'
+								onChange={handleChange}
+								placeholder='password'
+								value={password}
+								error={errors['password']}
+							/>
+							<InputField
+								type='password'
+								name='confirm_password'
+								label='Confirm Password'
+								onChange={confirmPass}
+								placeholder='confirm password'
+								value={cpassword}
+								error={(cpassword !== password) ? ['Password mismatch'] : []}
+							/>
+							<Button
+								type="submit"
+								className="btn btn-lg btn-success btn-block btn-signin"
+								text="Register"
+							/>
+							<p className="message">Alredy registered? <Link to="/login">Signin</Link></p>
+						</form>
 					</div>
 				</div>
 			</div>
-		);
-	}
-}
-
+		</div>
+	);
+};
+RegisterForm.propTypes = {
+	handleChange: PropTypes.func.isRequired,
+	handleSubmit: PropTypes.func.isRequired,
+	confirmPass: PropTypes.func.isRequired,
+	username: PropTypes.string,
+	password: PropTypes.string,
+	cpassword: PropTypes.string,
+	email: PropTypes.string,
+	first_name: PropTypes.string,
+	last_name: PropTypes.string,
+};
 export default RegisterForm;
