+++ src/Components/inputs/InputField.jsx
@@ -30,7 +30,7 @@ InputField.propTypes = {
 	onChange: PropTypes.func.isRequired,
 	placeholder: PropTypes.string,
 	value: PropTypes.string,
-	error: PropTypes.string,
+	error: PropTypes.array,
 	required: PropTypes.string,
 	disabled: PropTypes.string
 };
