+++ src/utils/Config.jsx
@@ -1,2 +1,2 @@
-export const baseURL = 'http://0.0.0.0:8081/api/v2';
+export const baseURL = 'https://weconnect1-api.herokuapp.com/api/v2';
 export const cloudName = 'dzmdvppit';
